Test with int

+++ src/environment/domain/FirstObservation.java
@@ -6,15 +6,15 @@ import java.util.List;
 import java.util.Map;
 
 public class FirstObservation {
-    private final List<Double> observationVector;
+    private final List<Integer> observationVector;
     private final Map<String, String> info;
 
-    public FirstObservation(List<Double> observationVector, Map<String, String> info) {
+    public FirstObservation(List<Integer> observationVector, Map<String, String> info) {
         this.observationVector = observationVector;
         this.info = info;
     }
 
-    public List<Double> getObservationVector() {
+    public List<Integer> getObservationVector() {
         return observationVector;
     }
